[3701] - introduce namespacing for course vacancies updated mailer variants
